leader-toolbox/pyproject.toml
[project]
name = "leader-toolbox"
version = "0.1.0"
description = "A Chat web application allowing to retrieve info from KB (Knowledge Base) and accomplish some leadership tasks."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "faiss-cpu==1.7.4",
    "fastapi>=0.95",
    "pytest>=7.0",
    "sentence-transformers>=2.2.2",
    "uvicorn[standard]>=0.22",
]
